crew/pyproject.toml
[project]
name = "crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.134.0",
    "crewai-tools>=0.48.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]
